# Benchmark JetStream consumer with filters