Creates a consumer with a number of filters pulling in order from a stream with high subject cardinality.
(Inspired by a similar [HoloPlot](https://github.com/holoplot) benchmark)


Signed-off-by: Marco Primi <marco@synadia.com>