@mprimi 2024-01-22
## Example result
Run benchmark before and after the [recent stree](https://github.com/nats-io/nats-server/pull/4983) change

```
❯❯❯ git rebase -i d9235abb
❯❯❯ go test ./server -test.bench '^BenchmarkJetStreamConsumeWithFilters$' -test.run '^$' -test.benchtime 3s -test.count 5 > before.stree.d9235abb.txt


❯❯❯ git rebase -i fd1e1e68
❯❯❯ go test ./server -test.bench '^BenchmarkJetStreamConsumeWithFilters$' -test.run '^$' -test.benchtime 3s -test.count 5 > after.stree.fd1e1e68.txt


❯❯❯ benchstat before.stree.d9235abb.txt after.stree.fd1e1e68.txt
goos: darwin
goarch: arm64
pkg: github.com/nats-io/nats-server/v2/server
                                                              │ before.stree.d9235abb.txt │      after.stree.fd1e1e68.txt       │
                                                              │          sec/op           │    sec/op     vs base               │
JetStreamConsumeWithFilters/N=1,R=1/D=100,DS=10,F=5,C=12-8                   800.1µ ± ∞ ¹   523.0µ ± ∞ ¹  -34.64% (p=0.008 n=5)
JetStreamConsumeWithFilters/N=1,R=1/D=1000,DS=10,F=25,C=12-8                 38.89m ± ∞ ¹   24.46m ± ∞ ¹  -37.10% (p=0.008 n=5)
JetStreamConsumeWithFilters/N=1,R=1/D=10000,DS=10,F=50,C=12-8                854.3m ± ∞ ¹   538.6m ± ∞ ¹  -36.96% (p=0.008 n=5)
JetStreamConsumeWithFilters/N=3,R=3/D=100,DS=10,F=5,C=12-8                   917.2µ ± ∞ ¹   632.0µ ± ∞ ¹  -31.10% (p=0.008 n=5)
JetStreamConsumeWithFilters/N=3,R=3/D=1000,DS=10,F=25,C=12-8                 38.39m ± ∞ ¹   24.66m ± ∞ ¹  -35.77% (p=0.008 n=5)
JetStreamConsumeWithFilters/N=3,R=3/D=10000,DS=10,F=50,C=12-8                881.9m ± ∞ ¹   539.7m ± ∞ ¹  -38.80% (p=0.008 n=5)
geomean                                                                      30.63m         19.67m        -35.77%

```