@svenfoo 2024-01-23
Interesting, and impressive improvements!

Our in production scenario, and also the [nats-bench](https://github.com/holoplot/nats-bench/blob/main/client/client.go#L143C22-L143C44), uses a file storage for the stream. So I guess that this won't have an effect there, is that correct?