@Jarema 2024-01-23
It should have the same positive effect on file based storage @svenfoo