@svenfoo 2024-01-23
> It should have the same positive effect on file based storage @svenfoo

Awesome, I'll test the nightly as soon as I can.