@mprimi 2024-01-24
This moved into https://github.com/nats-io/nats-server/pull/4997